[Security Solution] move some shared files to flyout_v2 folder
