+++ x-pack/solutions/security/plugins/security_solution/public/flyout/document_details/left/components/related_alerts_by_session.test.tsx
@@ -21,7 +21,7 @@ import {
   EXPANDABLE_PANEL_HEADER_TITLE_ICON_TEST_ID,
   EXPANDABLE_PANEL_HEADER_TITLE_TEXT_TEST_ID,
   EXPANDABLE_PANEL_TOGGLE_ICON_TEST_ID,
-} from '../../../shared/components/test_ids';
+} from '../../../../flyout_v2/shared/components/test_ids';
 
 jest.mock('../../shared/hooks/use_fetch_related_alerts_by_session');
 jest.mock('../hooks/use_paginated_alerts');
