+++ x-pack/solutions/security/plugins/security_solution/public/flyout/document_details/left/components/related_cases.tsx
@@ -19,7 +19,7 @@ import {
   CORRELATIONS_DETAILS_CASES_SECTION_TEST_ID,
 } from './test_ids';
 import { useFetchRelatedCases } from '../../shared/hooks/use_fetch_related_cases';
-import { ExpandablePanel } from '../../../shared/components/expandable_panel';
+import { ExpandablePanel } from '../../../../flyout_v2/shared/components/expandable_panel';
 
 const ICON = 'warning';
 const EXPAND_PROPERTIES = {
