+++ x-pack/solutions/security/plugins/security_solution/public/flyout/document_details/left/components/suppressed_alerts.test.tsx
@@ -17,7 +17,7 @@ import {
   EXPANDABLE_PANEL_HEADER_TITLE_ICON_TEST_ID,
   EXPANDABLE_PANEL_HEADER_TITLE_TEXT_TEST_ID,
   EXPANDABLE_PANEL_TOGGLE_ICON_TEST_ID,
-} from '../../../shared/components/test_ids';
+} from '../../../../flyout_v2/shared/components/test_ids';
 import { DocumentDetailsContext } from '../../shared/context';
 import { mockContextValue } from '../../shared/mocks/mock_context';
 import { isSuppressionRuleInGA } from '../../../../../common/detection_engine/utils';
