+++ x-pack/solutions/security/plugins/security_solution/public/flyout/document_details/right/components/entities_overview.test.tsx
@@ -24,7 +24,7 @@ import {
   EXPANDABLE_PANEL_HEADER_TITLE_LINK_TEST_ID,
   EXPANDABLE_PANEL_HEADER_TITLE_TEXT_TEST_ID,
   EXPANDABLE_PANEL_TOGGLE_ICON_TEST_ID,
-} from '../../../shared/components/test_ids';
+} from '../../../../flyout_v2/shared/components/test_ids';
 import { useRiskScore } from '../../../../entity_analytics/api/hooks/use_risk_score';
 import { useNavigateToLeftPanel } from '../../shared/hooks/use_navigate_to_left_panel';
