+++ x-pack/solutions/security/plugins/security_solution/public/flyout/document_details/right/components/investigation_section.tsx
@@ -7,17 +7,19 @@
 
 import React, { memo, useMemo } from 'react';
 import { EuiSpacer } from '@elastic/eui';
-import { FormattedMessage } from '@kbn/i18n-react';
 import { buildDataTableRecord, type DataTableRecord, type EsHitRecord } from '@kbn/discover-utils';
+import { FLYOUT_STORAGE_KEYS } from '../../../../flyout_v2/document/constants/local_storage';
 import { useExpandSection } from '../../../../flyout_v2/shared/hooks/use_expand_section';
 import { ExpandableSection } from '../../../../flyout_v2/shared/components/expandable_section';
 import { HighlightedFields } from './highlighted_fields';
-import { INVESTIGATION_SECTION_TEST_ID } from '../../../../flyout_v2/document/components/investigation_section';
+import {
+  INVESTIGATION_SECTION_TEST_ID,
+  INVESTIGATION_SECTION_TITLE,
+} from '../../../../flyout_v2/document/components/investigation_section';
 import { InvestigationGuide } from '../../../../flyout_v2/document/components/investigation_guide';
 import { getField } from '../../shared/utils';
 import { EventKind } from '../../shared/constants/event_kinds';
 import { useDocumentDetailsContext } from '../../shared/context';
-import { FLYOUT_STORAGE_KEYS } from '../../shared/constants/local_storage';
 import { useNavigateToLeftPanel } from '../../shared/hooks/use_navigate_to_left_panel';
 import { LeftPanelInvestigationTab } from '../../left';
 
@@ -58,12 +60,7 @@ export const InvestigationSection = memo(() => {
   return (
     <ExpandableSection
       expanded={expanded}
-      title={
-        <FormattedMessage
-          id="xpack.securitySolution.flyout.right.investigation.sectionTitle"
-          defaultMessage="Investigation"
-        />
-      }
+      title={INVESTIGATION_SECTION_TITLE}
       localStorageKey={FLYOUT_STORAGE_KEYS.OVERVIEW_TAB_EXPANDED_SECTIONS}
       sectionId={KEY}
       gutterSize="none"
