+++ x-pack/solutions/security/plugins/security_solution/public/flyout/document_details/right/components/session_preview_container.tsx
@@ -11,7 +11,7 @@ import { useLicense } from '../../../../common/hooks/use_license';
 import { SessionPreview } from './session_preview';
 import { useSessionViewConfig } from '../../shared/hooks/use_session_view_config';
 import { useDocumentDetailsContext } from '../../shared/context';
-import { ExpandablePanel } from '../../../shared/components/expandable_panel';
+import { ExpandablePanel } from '../../../../flyout_v2/shared/components/expandable_panel';
 import { SESSION_PREVIEW_TEST_ID } from './test_ids';
 import { useNavigateToSessionView } from '../../shared/hooks/use_navigate_to_session_view';
 import { SessionViewNoDataMessage } from '../../shared/components/session_view_no_data_message';
