+++ x-pack/solutions/security/plugins/security_solution/public/flyout/entity_details/user_right/components/managed_user_accordion.tsx
@@ -12,7 +12,7 @@ import { FormattedMessage } from '@kbn/i18n-react';
 import { get } from 'lodash/fp';
 import type { EntityDetailsPath } from '../../shared/components/left_panel/left_panel_header';
 import { EntityDetailsLeftPanelTab } from '../../shared/components/left_panel/left_panel_header';
-import { ExpandablePanel } from '../../../shared/components/expandable_panel';
+import { ExpandablePanel } from '../../../../flyout_v2/shared/components/expandable_panel';
 import type { ManagedUserFields } from '../../../../../common/search_strategy/security_solution/users/managed_details';
 import { FormattedRelativePreferenceDate } from '../../../../common/components/formatted_date';
 import { ONE_WEEK_IN_HOURS } from '../../shared/constants';
