+++ x-pack/solutions/security/plugins/security_solution/public/flyout/shared/components/test_ids.ts
@@ -12,20 +12,6 @@ export const FLYOUT_LINK_TEST_ID = `${PREFIX}Link` as const;
 
 export const FLYOUT_ERROR_TEST_ID = `${PREFIX}Error` as const;
 export const FLYOUT_LOADING_TEST_ID = `${PREFIX}Loading` as const;
-export const EXPANDABLE_PANEL_TOGGLE_ICON_TEST_ID = (dataTestSubj: string) =>
-  `${dataTestSubj}ToggleIcon`;
-export const EXPANDABLE_PANEL_HEADER_LEFT_SECTION_TEST_ID = (dataTestSubj: string) =>
-  `${dataTestSubj}LeftSection`;
-export const EXPANDABLE_PANEL_HEADER_TITLE_ICON_TEST_ID = (dataTestSubj: string) =>
-  `${dataTestSubj}TitleIcon`;
-export const EXPANDABLE_PANEL_HEADER_TITLE_LINK_TEST_ID = (dataTestSubj: string) =>
-  `${dataTestSubj}TitleLink`;
-export const EXPANDABLE_PANEL_HEADER_TITLE_TEXT_TEST_ID = (dataTestSubj: string) =>
-  `${dataTestSubj}TitleText`;
-export const EXPANDABLE_PANEL_HEADER_RIGHT_SECTION_TEST_ID = (dataTestSubj: string) =>
-  `${dataTestSubj}RightSection`;
-export const EXPANDABLE_PANEL_LOADING_TEST_ID = (dataTestSubj: string) => `${dataTestSubj}Loading`;
-export const EXPANDABLE_PANEL_CONTENT_TEST_ID = (dataTestSubj: string) => `${dataTestSubj}Content`;
 
 /* Header Navigation */
