+++ x-pack/solutions/security/plugins/security_solution/public/flyout_v2/document/constants/local_storage.ts
@@ -1,0 +1,11 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0; you may not use this file except in compliance with the Elastic License
+ * 2.0.
+ */
+
+export const FLYOUT_STORAGE_KEYS = {
+  OVERVIEW_TAB_EXPANDED_SECTIONS:
+    'securitySolution.documentDetailsFlyout.overviewSectionExpanded.v8.14',
+};
